generation/pyproject.toml
[tool.poetry]
name = "chugunapps"
version = "0.1.0"
description = "my portfolio (backend development)"
authors = ["chugunova24 <t.chugunova24@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
asgiref = "==3.7.2"
attrs = "==23.1.0"
autopep8 = "==2.0.4"
build = "==0.10.0"
cachecontrol = "==0.13.1"
certifi = "==2023.7.22"
cffi = "==1.16.0"
charset-normalizer = "==3.3.0"
cleo = "==2.0.1"
crashtest = "==0.4.1"
crispy-bootstrap4 = "==2022.1"
cryptography = "==41.0.4"
distlib = "==0.3.7"
django = "==4.1"
django-crispy-forms = "==2.0"
django-debug-toolbar = "==4.2.0"
django-extensions = "==3.2.3"
django-filter = "==23.2"
django-phonenumber-field = "==7.1.0"
django-redis = "==5.3.0"
django-silk = "==5.0.4"
djangorestframework = "==3.14.0"
dulwich = "==0.21.6"
filelock = "==3.12.4"
gprof2dot = "==2022.7.29"
html2text = "==2020.1.16"
idna = "==3.4"
importlib-metadata = "==6.8.0"
installer = "==0.7.0"
jaraco-classes = "==3.3.0"
jeepney = "==0.8.0"
jsonschema = "==4.17.3"
keyring = "==24.2.0"
markdown = "==3.4.3"
more-itertools = "==10.1.0"
msgpack = "==1.0.7"
numpy = "==1.25.1"
packaging = "==23.1"
pandas = "==2.0.3"
pexpect = "==4.8.0"
phonenumbers = "==8.13.18"
pillow = "==10.0.0"
pkginfo = "==1.9.6"
platformdirs = "==3.10.0"
poetry = "==1.6.1"
poetry-core = "==1.7.0"
poetry-plugin-export = "==1.5.0"
ptyprocess = "==0.7.0"
pycodestyle = "==2.11.0"
pycparser = "==2.21"
pyproject-hooks = "==1.0.0"
pyrsistent = "==0.19.3"
python-dateutil = "==2.8.2"
python-dotenv = "==1.0.0"
pytz = "==2023.3"
rapidfuzz = "==2.15.1"
redis = "==5.0.0"
requests = "==2.31.0"
requests-toolbelt = "==1.0.0"
secretstorage = "==3.3.3"
shellingham = "==1.5.3"
six = "==1.16.0"
sqlparse = "==0.4.4"
tomlkit = "==0.12.1"
trove-classifiers = "==2023.9.19"
tzdata = "==2023.3"
urllib3 = "==2.0.5"
virtualenv = "==20.24.5"
zipp = "==3.17.0"
psycopg2-binary = "^2.9.6"


[build-system]
requires = ["poetry-core>=$POETRY_VERSION"]
build-backend = "poetry.core.masonry.api"
